services/pg_tx_fetcher/requirements.txt
asttokens==2.4.1
certifi==2024.2.2
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
frozendict==2.3.10
jedi==0.19.1
matplotlib-inline==0.1.7
moralis==0.1.46
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.3.0
urllib3==1.26.18
wcwidth==0.2.13
